Created class cLassifierTrainer.java Added naivebayes and randomforest model trainer
